+++ src/components/GameDetailView/guards/index.ts
@@ -1,0 +1,6 @@
+export * from "./hasDuration";
+export * from "./hasGenres";
+export * from "./hasHltbCompletionist";
+export * from "./hasHltbExtra";
+export * from "./hasHltbMain";
+export * from "./hasReleaseDate";
